feat (basket2-3) CGB-63 I adjusted the code for the API in advance

+++ FakeNFT/Scenes/Basket/Cart/BacketViewController.swift
@@ -119,7 +119,7 @@ final class BacketViewController: UIViewController, BacketViewProtocol {
         nftCountLabel.text = "\(count) NFT"
     }
 
-    func updateTotalPriceLabel(with totalPrice: Double) {
+    func updateTotalPriceLabel(with totalPrice: Float) {
         priceLabel.text = String(format: "%.2f ETH", totalPrice)
     }
